Add dedicated card layouts (#884)
* Add dedicated layout for tuition fees card

The tuition fees card previously used the general card_item.xml. This
apparently resulted in a crash when a dynamically added TextView was not
found.

To fix this, the tuition fees card now provides a dedicated layout
card_tuition_fees.xml with all its required views.

* Add dedicated layouts and ViewHolders for Cards that used card_item.xml

This is a first step towards making the existing way of populating ViewHolders easier and less error-prone. No longer uses card_item.xml, but dedicated layouts for each card type.

Also, add subclasses of CardViewHolder which are responsible for binding data to a card’s view.

* Add didBind flag to prevent content from being added multiple times

The content of MVVCard, ChatMessagesCard and CafeteriaMenuCard are added dynamically. To prevent the content from being added every time the ViewHolder becomes visible, I added didBind as a boolean variable.

* Switch to CafeteriaWithMenus in CafeteriaMenuCard

* Remove commented code