Move DAO instantiation to onCreate to prevent NullPointerException (#1001)
Do not use "this" as a Context in the constructor of an Activity. In TcaDb, we call context.getApplicationContext(), which will eventually invoke ContextWrapper.getApplicationContext(). There, getApplicationContext() is called on its property mBase, which is a Context. However, mBase is only set with attachBaseContext(). If we use "this" in the constructor, this method has not run yet.