Make improvements to cards (#1100)
* Add “more" icon to cards

* Refactor NewsCard adapter for easier distinction between news and film items

* Remove commented code

* Increase tappable area of options icon

To do this, I had to move to ConstraintLayout for many card layouts. Otherwise, the increased height of the options icon would have resulted in an odd additional vertical offset below the title.

* Improve "Always hide card" functionality

Now, cards are immediately removed when "Always hide card" is selected

* Add additional padding around options button in news and event cards

This increases the tap area, making the button easier to hit.

* Only offer "Open card settings" for cards that actually have settings

* Add card options to login card

* Remove commented code in layout files

* Remove commented code

* Add clarifying comment regarding chat messages view inflation

* Change how cards declare the options menu that they want to have displayed

* Fix bug related to displaying card options

Remember, kids: Copy-and-paste programming is bad.