Finish hide buy button PR (#1311)
* Implemented measures to hide >Buy Ticket< Button and number of available Tickets if the number of available Tickets is negative.
This is especially usefull for events that should be displayed as such, but do not sell tickets via the app (e.g. Diversityfilm, UNITY, MaiTUM, TUNIX, GARNIX, ...).
Additionally, when no more online tickets are available, the number 0 is now replaced with a text explaining that tickets may still be available at the evening box office.

* Fixed Bug that prevented user from viewing purchased tickets if tickets were sold out.

* Switched to official spelling >der tu film<.

* Replaced ticket string with plural.

* Implemented measures to hide >Buy Ticket< Button and number of available Tickets if the number of available Tickets is negative.
This is especially usefull for events that should be displayed as such, but do not sell tickets via the app (e.g. Diversityfilm, UNITY, MaiTUM, TUNIX, GARNIX, ...).
Additionally, when no more online tickets are available, the number 0 is now replaced with a text explaining that tickets may still be available at the evening box office.

* Fixed Bug that prevented user from viewing purchased tickets if tickets were sold out.

* Switched to official spelling >der tu film<.

* Replaced ticket string with plural.

* use ticket status to determine if we offer tickets at all and rename function

* put conditions into functions

* add comment to keep code the same and remove repetitiveness/readability

* fix comment

* fix bug